backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify-client>=1.8.1",
    "beautifulsoup4>=4.14.3",
    "crawl4ai>=0.4.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.26.0",
    "pdfplumber>=0.11.9",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "selenium>=4.41.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
